feat: refactor layout and filter functionality in index.html and scripts
- Updated index.html to use new CSS files for styles and split view.
- Implemented a filter toggle button for better UI/UX in the filters section.
- Removed deprecated split-view.js and integrated its functionality into the new split-view.js.
- Enhanced responsiveness and layout management for both desktop and mobile views.
