+++ css/split-view.css
