+++ css/styles.css
@@ -1,0 +1,293 @@
+/* Reset and base styles */
+* {
+  margin: 0;
+  padding: 0;
+  box-sizing: border-box;
+}
+
+html,
+body {
+  margin: 0;
+  padding: 0;
+  font-family: Arial, sans-serif;
+  overflow: hidden;
+  height: 100%;
+  width: 100%;
+}
+
+/* Estilos para el encabezado */
+header {
+  background-color: #4CAF50;
+  color: white;
+  padding: 10px 20px;
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  box-shadow: 0 2px 5px rgba(0, 0, 0, 0.1);
+}
+
+header h1 {
+  font-size: 1.5rem;
+  margin: 0;
+}
+
+.header-controls {
+  display: flex;
+  align-items: center;
+  gap: 15px;
+}
+
+.info-button {
+  background: none;
+  border: none;
+  color: white;
+  cursor: pointer;
+  padding: 5px;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
+
+/* Estilos para los paneles */
+.panel {
+  overflow: auto;
+  padding: 20px;
+  box-sizing: border-box;
+}
+
+/* Estilos para los filtros */
+.filters {
+  margin-bottom: 20px;
+}
+
+.filters-row {
+  display: flex;
+  flex-wrap: wrap;
+  gap: 15px;
+  margin-bottom: 10px;
+}
+
+.search-field, .city-search, .type-filter {
+  flex: 1;
+  min-width: 200px;
+}
+
+.input-with-icon, .select-container {
+  position: relative;
+  display: flex;
+  align-items: center;
+}
+
+.input-with-icon input, .select-container select {
+  width: 100%;
+  padding: 8px 30px 8px 10px;
+  border: 1px solid #ddd;
+  border-radius: 4px;
+  font-size: 14px;
+}
+
+.search-icon {
+  position: absolute;
+  right: 30px;
+  color: #4CAF50;
+}
+
+.clear-input {
+  position: absolute;
+  right: 5px;
+  background: none;
+  border: none;
+  cursor: pointer;
+  padding: 5px;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
+
+/* Estilos para la lista */
+.list-container {
+  overflow-y: auto;
+  max-height: calc(100vh - 250px);
+}
+
+#establishment-list {
+  list-style: none;
+  padding: 0;
+}
+
+#establishment-list li {
+  padding: 10px;
+  border-bottom: 1px solid #eee;
+  cursor: pointer;
+  transition: background-color 0.2s;
+}
+
+#establishment-list li:hover {
+  background-color: #f5f5f5;
+}
+
+#establishment-list li.active {
+  background-color: #e8f5e9;
+  border-left: 3px solid #4CAF50;
+}
+
+.no-results {
+  text-align: center;
+  color: #666;
+  padding: 20px;
+}
+
+.result-count {
+  font-size: 0.9rem;
+  color: #666;
+  padding: 5px 10px;
+  background-color: #f5f5f5;
+  border-radius: 4px;
+  margin-bottom: 10px;
+}
+
+/* Estilos para el mapa */
+#map {
+  height: calc(100vh - 150px);
+  width: 100%;
+  border-radius: 4px;
+}
+
+/* Estilos para el modal */
+.modal {
+  display: none;
+  position: fixed;
+  z-index: 1000;
+  left: 0;
+  top: 0;
+  width: 100%;
+  height: 100%;
+  background-color: rgba(0, 0, 0, 0.5);
+}
+
+.modal-content {
+  background-color: #fff;
+  margin: 10% auto;
+  padding: 20px;
+  border-radius: 5px;
+  width: 80%;
+  max-width: 500px;
+  position: relative;
+}
+
+.close-modal {
+  position: absolute;
+  right: 10px;
+  top: 5px;
+  font-size: 24px;
+  cursor: pointer;
+}
+
+.info-content {
+  margin-top: 15px;
+}
+
+.info-content p {
+  margin-bottom: 10px;
+}
+
+/* Estilos para los botones de control de vista */
+.view-controls {
+  display: flex;
+  gap: 5px;
+}
+
+.view-control-btn {
+  background-color: rgba(255, 255, 255, 0.2);
+  border: 1px solid rgba(255, 255, 255, 0.3);
+  color: white;
+  padding: 5px 10px;
+  cursor: pointer;
+  border-radius: 4px;
+  transition: background-color 0.2s;
+  font-size: 0.9rem;
+}
+
+.view-control-btn:hover {
+  background-color: rgba(255, 255, 255, 0.3);
+}
+
+.view-control-btn.active {
+  background-color: white;
+  color: #4CAF50;
+  border-color: white;
+}
+
+/* Estilos para dispositivos móviles */
+@media (max-width: 768px) {
+  header {
+      flex-direction: column;
+      align-items: flex-start;
+      gap: 10px;
+  }
+  
+  .header-controls {
+      width: 100%;
+      justify-content: space-between;
+  }
+  
+  header h1 {
+      font-size: 1.2rem;
+  }
+  
+  .panel {
+      padding: 10px;
+  }
+  
+  #map {
+      height: calc(100vh - 200px);
+  }
+  
+  .view-controls {
+      flex-wrap: wrap;
+  }
+  
+  .view-control-btn {
+      font-size: 0.8rem;
+      padding: 4px 8px;
+  }
+  
+  .filters-row {
+      flex-direction: column;
+      gap: 10px;
+  }
+  
+  .search-field, .city-search, .type-filter {
+      width: 100%;
+  }
+}
+
+.filters-toggle {
+  margin-bottom: 1rem;
+  text-align: right;
+}
+
+.toggle-button {
+  background-color: #4CAF50;
+  color: white;
+  border: none;
+  padding: 8px 16px;
+  border-radius: 4px;
+  cursor: pointer;
+  display: inline-flex;
+  align-items: center;
+  gap: 8px;
+  transition: background-color 0.3s;
+}
+
+.toggle-button:hover {
+  background-color: #45a049;
+}
+
+.toggle-icon {
+  transition: transform 0.3s;
+}
+
+.toggle-button.active .toggle-icon {
+  transform: rotate(180deg);
+}
